+++ packages/keyring-api/CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add `TransactionType.TokenDisapprove` for token disapproval transactions and optional `typeLabel` string field to `TransactionDetails` ([#568](https://github.com/MetaMask/accounts/pull/568))
+- Add `TransactionType.TokenDisapprove` for token disapproval transactions ([#568](https://github.com/MetaMask/accounts/pull/568))
+- Add optional `TransactionDetails.typeLabel` ([#568](https://github.com/MetaMask/accounts/pull/568))
+  - This can be used to display a custom label regarding the transaction type.
 
 ## [23.2.0]